First pass at the addition of a REST API.
Contains a new restapi subproject, with a single controller which
receives input requests and dispatches them to an endpoint as valid hl7
input messages.

Also contains updated tests and inputs for the common package designed
to receive the requests sent by the REST API and split them into batch
messages for processing.